Calculate loan application amount

+++ src/main/java/chamapool/application/loans/responses/LoanEligibilityResponse.java
@@ -1,3 +1,3 @@
 package chamapool.application.loans.responses;
 
-public record LoanEligibilityResponse(boolean isEligible, Integer amountEligible, String reason) {}
+public record LoanEligibilityResponse(boolean isEligible, Double amountEligible, String reason) {}
